codegen: use `rsp` instead of `sp` in stack frame setup code
`sp` is the 16-bit version of the stack pointer `rsp`, and arithmetic on
it leaves the upper 48 bit of `rsp` unchanged. This mostly works, but
has a 64k / N (where N is the size of the stack frame) chance of
failing: when `sub rsp, N` crosses a 64k boundary, `sub sp, N` will be
off by 64k and the first stack access will be out-of-bounds and produce
a segfault.